+++ GameKeyCasino/GameCasino.Dao.Interfaces/IUserDao.cs
@@ -5,8 +5,8 @@ namespace GameCasino.Dao.Interfaces
     public interface IUserDao
     {
         void Add(User user);
-        void AddMoney(int idUser,float money);
-        void RemoveMoney(int idUser,float money);
+        void AddMoney(int idUser,decimal money);
+        void RemoveMoney(int idUser,decimal money);
 
     }
 }
